Error on multiple __proto__ properties in object types
Summary:
Like __proto__ properties on object literal values, we special-case this syntax
to specify an object type with a specific prototype other than Object.prototype.
However, it's unclear what behavior we should have when multiple such properties
are defined in a single object type.

Before this change, we would simply take the last one, ignoring any that came
before. After this change, we take the first one and error on (and ignore) any
subsequent ones.

I doubt that this will cause any errors in practice, but the erroring behavior
seems much more explicit and helpful. This is also consistent with multiple
indexers defined on an object type.

Reviewed By: jbrown215

Differential Revision: D14677197

fbshipit-source-id: dc716072dde3a4aa300e3a9a978f108e3da03c85