Include subtimers in scuba data
Summary:
Our profiling is hierarchical, but we flatten it out for scuba. Currently we include:

1. The top level timer (which is a single total time). This is always renamed to `Profiling`
2. The subtimers of the top level timer.

So given something like

```
<Total>
 FetchSavedState
 LoadSavedState
   SavedStateLoad
   SavedStateDecompress
   SavedStateDenormalize
   <Unknown total>
 RestoreHeaps
<Unknown total>
```

We wouldn't log the `SavedStateLoad`, `SavedStateDecompress`, or `SavedStateDenormalize` timers to scuba. We'd have to choose whether we want a "total time to load saved state" timer, like `LoadSavedState`, or whether to log the granular timers at the top level.

So I want to change the logging to now include

1. The top level timer (which is a single total time). This is always renamed to `Profiling`
2. The subtimers of the top level timer.
3. New: The subtimers of the subtimes.

So something like

```
<Total>
 FetchSavedState
 LoadSavedState
   Read
   Decompress
   Denormalize
   <Unknown total>
 RestoreHeaps
<Unknown total>
```

Would log `Profiling`, `FetchSavedState`, `LoadSavedState`, `LoadSavedState:Read`, `LoadSavedState:Decompress`, `LoadSavedState:Denormalize`, and `RestoreHeaps`

Reviewed By: samwgoldman

Differential Revision: D14989035

fbshipit-source-id: 3c74f80298efae67429f43a916303a1bd391f688