Speed up saved state loading by delaying normalization
Summary:
This significantly speeds up loading saved states.

A major part of loading a saved state is denormalizing the data (turning relative paths back into absolute paths). This was very slow, so I looked into why it was slow. The current implementation was slow because we had to write and read a lot of data to the workers. But it was a little too much work to without the workers.

The solution was to delay normalizing this data. Later, during the `RestoreHeaps` step, we were using workers to restore the data to the heaps anyway. We can just denormalize the data then, and save a worker roundtrip.

Reviewed By: samwgoldman

Differential Revision: D15034335

fbshipit-source-id: 7b1912706ee278012dd2226ac1fd2db6eeaa3477