[hack] simplify hhi dependency
Summary:
In order to build the src/hhi submodule, we need to depend on the hhi toplevel
directory. This patch is an attempt to make it easier and to actually have
the hhi files in the _build directory. This is motivated by the coroutine
test suits that also needs these hhi files

Reviewed By: andrewjkennedy

Differential Revision: D15061374

fbshipit-source-id: 99d7d945978f34201c67f44ab34659eaf0af8241