[hack] Write multi-file test cases to disk during testing
Summary: Similar to the previous commit.  We have some test cases which are declared in a single file, but which define multiple files using special inline comments. We need to actually write those files to disk so that we can compute decls on-demand by parsing them when our caches have zero capacity.

Reviewed By: pittsw

Differential Revision: D14936033

fbshipit-source-id: d45af4a402c64cd428de28ffbbbd3bd2b707223a