Make function type statics stricter
Summary:
Before this change, we optimistically typed the statics of a function type as
`any`, under the assumption that a function would not be used as an object. In
practice, our optimism here is unfounded.

This diff changes the type of the statics object for function types to an empty,
inexact object type with Function.prototype for its __proto__. This choice is
sound with respect to the runtime behaviors while also allowing functions with
statics to be valid subtypes via width subtyping.

Reviewed By: panagosg7

Differential Revision: D14448564

fbshipit-source-id: 87f6bf50a795760023e1e237856e80587df9f857