[react] normalize configs via object_kit instead of CondT
Summary:
The purpose of the CondT here is to turn null or undefined into `{||}`, and do nothing to any other type. Unfortunately, that doesn't work so well with (at least) intersection types. See the discussion in D14578158 for more details about that.

The reason we need to turn null or undefined into `{||}` is so that we can lookup the ref and key in the config. Instead of using CondT, we can use object_kit to get an object representation of the config type (which has cases for null, empty, undefined, mixed, etc.) and perform the lookup there instead.

This change reduces the err diff on a fix for function components (later in the stack) from ~3600 to ~200 errors.

Reviewed By: samwgoldman

Differential Revision: D14686167

fbshipit-source-id: 10d3618fef1032794df58922c39e54ceaa92ce4e