[react] Use ConfigCheck to check function component configs
Summary:
Previously, we would read the config from the function component to perform the check. However, function components don't require annotations, so that config was never resolved if the props didn't have a lower bound.

Instead, we should model this as trying to call createElement on the component using the given config. This doesn't require a lower bound on props to fire. See the added test for an example of an error we used to miss.

I am less than satisfied with the error messages on unannotated components. For example, see the lazy.js test in react_16_6. Since the tvar for the Props in lazy receives lower bounds from the createElement calls and an upper bound from the input component, we end up comparing those object types on the tvar and error with the primary location set to the call of lazy (since that's where the tvar is instantiated). I spent some time trying to improve this, but had no luck.

Reviewed By: samwgoldman

Differential Revision: D14566404

fbshipit-source-id: 4c2dfd9a8591837b95ddfb16cdd88ba22e6fb599