[react] React.ConfigCheck to determine config compatibility
Summary:
The motivation behind this change is to fix an issue where unannotated function components can be unsoundly casted to an AbstractComponent with any config type.

To fix this, we plan to model the flow as calling createElement on the component with the config type specified by the AbstractComponent.

create_element does a few things:

1. Checks the ref type is correct
2. Checks the key type is correct
3. Checks the config type is correct

We only need to do 3. to fix this bug.

2. causes spurious errors in HOCs, where the type of the config may be mixed, which is treated as {[string]: mixed}. When we check that `key` (if specified) is number | string, we instead hit the indexer and see mixed, which leads to an error.

3. Is already handled by the FunT ~> AbstractComponent flow. We could do 3 in create_element as well, but I didn't want to mess with inexact/exact objects intersecting with {ref: void}.

Reviewed By: samwgoldman

Differential Revision: D14566402

fbshipit-source-id: fe7655acc744d871553a3cd16e3a7c417ee3b043