[hack] Back out D14936033 "Write multi-file test cases to disk during testing"
Summary: The backed-out commit changed the signature of the "method jumps" RPC call to return a `Pos.t` instead of a `Pos.absolute`. `Pos.t` uses `Relative_path.t` internally. Relative paths must not be passed across the RPC boundary, because they can only be interpreted with the context of the repo root, which the client does not actually have, but the server does. Previously, the server would always send absolute paths over the RPC layer, which worked.

Reviewed By: andrewjkennedy

Differential Revision: D15193973

fbshipit-source-id: 9a9720ebccdb527cb580cf925a00e42890d30e32