don't expose the internal type language in extension error messages
Summary:
In certain contrived cases it was possible to get Flow to produce an error referencing an internal type representation rather than one that is actionable by the user. This produces a more sensible error message.

Note that the new test uses the `invariant` function in order to produce the error message, as Flow currently special-cases this name to perform refinements, even when this is not the actual definition of the function. This bug is separate from this change though, this was just the only way I could reproduce the behavior.

Reviewed By: avikchaudhuri

Differential Revision: D15253384

fbshipit-source-id: 54a95ded7c0963ce133164dac7fb0ecd78fd22d5