Change DestructuringT a use type instead of an eval type
Summary:
Now that the destructuring pass happens aftere substitution, it is no longer
necessary to use an eval type. This diff promotes the DestructuringT eval type
to a use type, which can be eagerly evaluated.

So, a nice simplification, but this change also sets the stage for fixing a
long-standing bug with union types and destructuring defaults.

This change does inadvertently change some reason descriptions in error
messages. I did not investigate on the source of the discrepency because the new
descriptions seemed at least as good if not better than before.

Reviewed By: avikchaudhuri

Differential Revision: D15272070

fbshipit-source-id: 2e82f24c80524c331910793ab5db97de00a6ba1f