propagating opts from merge phase to check phase of types-first
Summary:
In types-first, the merge phase is repurposed to merge dependencies. Some of those dependencies discover that their signatures have not changed. However, this information was not previously propagated to the (new) check phase, which caused all files that were merged to be checked again. (To add insult to injury, we were actually also checking unchanged but previously checked files again, but count that in the clowniness bucket.)

This diff fixes this situation. In particular, it collects information about which files had their signatures unchanged from Merge_stream, and uses this information to filter out dependents whose dependencies have not changed their signatures. This has to be done carefully so that errors for the skipped files are preserved.

Reviewed By: nmote

Differential Revision: D15197843

fbshipit-source-id: f979db1358f7ada2bba5cb65e27ab5daa37ef88f