[PR] HTMLElement: add options parameter for focus method
Summary:
Issue #7651: The HTMLElement focus method is missing the optional options parameter.
[MDN documentation on focusOptions parameter](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/focus#Syntax)

I added a new type for the focus options and added an optional parameter to the focus method on HTMLElement.

*HTMLInputElement*
I noticed that HTMLInputElement also has a focus() method on it, but I am unsure if HTMLInputElement's focus also accepts options. When I looked at the HTMLInputElement MDN docs, I was under the impression that it would also accept the same options since the focus() hyperlink directs you to the docs on HTMLElement.focus. I did not dig too much into this since I wanted to focus (😄) on a small change and understand how the contribution process works.

Please let me know if I am missing any steps or if anything is sorely incorrect... this is my first attempt at contributing. 😄
Pull Request resolved: https://github.com/facebook/flow/pull/7711

Reviewed By: jbrown215

Differential Revision: D15320106

Pulled By: dsainati1

fbshipit-source-id: 95901fe27a76faa5ed8c8189d50e12fe674b8dac