Deprecate `$Enum<...>` utility (in favour of `$Keys<...>`)
Summary:
`$Enum` is an outdated alias for `$Keys`, deprecate it. It is now a lint which by default is an error. Projects can temporarily not make it an error, but the plan is to completely remove it in the release after. We are aggressive since it can be replaced with a simple textual search/replace.

It is perhaps unfortunate that we have both `DeprecatedUtility` and `DeprecatedEnumUtility`, but `DeprecatedEnumUtility` will be completely removed soon so that won't be an issue for long.

Reviewed By: samwgoldman

Differential Revision: D15367312

fbshipit-source-id: 78d7a501fe88e1855508c14eebb9c4d60cc03981