Create both `Abstract` and `Concrete` variants of `ALoc.t`
Summary:
This is the next step in the abstract location project. We need to be able to store locations which have a concrete underlying representation as well as those with an abstract underlying representation.

Basically, all locations will start out as concrete but those returned by the signature builder will be converted to abstract locations.

My first attempt at an implementation led to significant wall time and memory regressions. This approach allows zero-cost conversions between `Loc.t` and (concrete) `ALoc.t`.

Reviewed By: jbrown215

Differential Revision: D13091131

fbshipit-source-id: 155e981ec97dc760a2c3a70f184dbf04ee59bf93