name the arguments to ESpeculationAmbiguous
Summary:
ESpeculationAmbiguous has 5 params that are usually unnamed. took a stab at giving them names so it's a bit more self-documenting.

not sure if `reason` and `reason_op` are entirely correct, though it's what we use in many other types. the callsite that creates an ESpeculationAmbiguous error calls them `case_r` and `r`, respectively.

Reviewed By: jbrown215

Differential Revision: D15358239

fbshipit-source-id: f0178ff704f9b86aab16aedffae234bc6d14fa63