[React] strictify `React$ComponentType` and `React$ElementType`
Summary:
This takes advantage of the fact that the `Config` type argument to `React$AbstractComponent` is contravariant, and the `Instance` argument is covariant to replace some important `any` types in the react libdefs.

This should have the additional benefit of allowing people to use `React.Element<React.ElementType>` in strict mode instead of `React.Element<any>`.

Reviewed By: jbrown215

Differential Revision: D14549278

fbshipit-source-id: 7043d3f9197eeaa25b343a507f7588561acae24f