[React] introduce React$MixedElement
Summary: Now that we have removed the `any` types from `React$ElementType` and `React$ComponentType`, it becomes feasible to produce a safe supertype of all React elements: `React$Element<React$ElementType>`. This introduces `React$MixedElement` as an alias for this supertype. The name was chosen to signify that this type behaves similarly to `mixed`, in that all elements are valid subtypes of it, but the data you can read off of it is limited.

Reviewed By: jbrown215

Differential Revision: D14549694

fbshipit-source-id: 9509b6bc8b9ecf0c6d7e872bf82e3a65a76e690a