fix `flow start --include-suppressed`
Summary:
the `--include-suppressed` flag returns suppressed errors when passed to `flow check`, but did nothing when passed to `flow start`.

aside: it has to be passed when starting the server, since the workers throw away suppressed errors if it's not set, so you can't do `flow status --include-suppressed`. perhaps `flow start --include-suppressed` should make the server remember all suppressed errors, and you should have to do `flow status --include-suppressed` to see them (or not). for most projects, the number of suppressions should be small enough for the server to remember them all.

Reviewed By: samwgoldman

Differential Revision: D15402401

fbshipit-source-id: 23468a85d18c89a8e234e5fd2ab0d32c653ac8ea