Make Object methods not take null or undefined
Summary:
Many methods like `Object.entries(obj)` take an argument which they cast to an object. Everything except `null` and `undefined` can be cast to an object. Casting `null` and `undefined` to an object is a runtime error, so Flow should error first.

We absolutely can expand on this in the future, and ban other conversions out of principle (like casting arrays to objects). But we definitely should ban converting `null` or `undefined` to objects now.

Reviewed By: jbrown215

Differential Revision: D15203273

fbshipit-source-id: b35a0cd13c28051d2a28ff0c10da922a08ab97dc