[trusted-types] Trust data in types can be either a trust status or an index for inference, with compact representation
Summary:
Currently, all `DefT`s contain trust information, but we assume that the trust of a type is fully known. This isn't where we want to end up, though: we'd love to actually *infer* whether a given type is trusted or untrusted, unless the user specifies that it must be trusted.

To do this inference, we need to be able to track interactions between types. For example, if `any` flows into `number`, then that `number` needs to be marked as untrusted (aka tainted). We'll accomplish this by reusing some of the ideas behind Flow's tvar inference, but in order to do so, the type whose trust is being inferred (in this case `number`) needs to have an index into a graph of trust inference variables, just like tvars contain an index into the context graph.

This diff alters the information contained within `DefT`s to be either a representation of trust (in the cases where the trust of a type is known or user-specified) or an index into a trust graph (currently unimplemented). The `trust.ml` module now provides two primary types: the `trust` type, as before, represents fully-known trust information, and the `trustdata` type contains either a `trust` value or an index into the graph. The `trustdata` type is now what appears in `DefT`s, allowing them to be either fixed trust or inferred trust.

The specifics of `trustdata` is complicated due to performance issues. What we'd really like is for the `trustdata` type to look like
```
type trustdata =
| Inferred of ident
| Fixed of trust
```
In other words, it should be either an ident or a trust value, and (importantly) client modules should be able to tell which is which--this should be an externally visible distinction. Unfortunately, since all `DefT`s contain a `trustdata` value, this results in a huge memory hit. As a result, we're now using a much more compact representation: a value of type `trustdata` is an int with the following layout:

```
trustdata:
[ tag | data ...... ]
  ^     ^
1 bit   62 bits
```
If `tag` is 0, then `data` should be interpreted as trust information. Trust information is now itself represented as an int, with values 0 to 3 representing different trust info (trusted public, untrusted private, etc).

If `tag` is 1, then `data` should be interpreted as an identifier, which (in future diffs) will be a pointer into a graph for trust inference.

This is all internal implementation details within `trust.ml`, of course. External modules use an API to manipulate `trustdata` values. The `expand` function takes a `trustdata` value and expands it to our ideal representation of `trustdata` as above, a datatype of either `Inferred` or `Trusted`. There are also predicates `is_infer` and `is_trust`, and a `trustdata` value can be directly converted to a `trust` value with `as_trust` and likewise for `as_infer` for inference identifiers (these functions assert false if the `trustdata` value is not actually a `trust` value or inference ident, respectively).

Reviewed By: dsainati1

Differential Revision: D14605940

fbshipit-source-id: ad078f9bf2a029fe6d370a43918eb92bcfc5165a