Delete incorrect enum case for SentinelPropTestT
Summary:
When flowing an object to a SentinelPropTestT with a union of enums as its test fields, we were previously requiring that all of the sentinel tests passed. This is not the right behavior for a union; when a union appears as an upper bound only one of the cases needs to pass.

Considering how rarely this case is hit, it doesn't seem necessary to build out a new speculation mode for this process, so just not refining and defaulting to the standard case is sufficient.

Reviewed By: avikchaudhuri

Differential Revision: D15383777

fbshipit-source-id: 108e636dcd941fc43982abce3d8de4215df8bc4e