[Perf] don't split unions for equality checks between unions
Summary: When testing equality of two unions (`==`, not `===`), we would previously split the union and check that every element of one union was equatable with every element of the other. Instead we can add a case to the equatability check for enums to skip the long quadratic check if the two unions are equal.

Reviewed By: avikchaudhuri

Differential Revision: D15355999

fbshipit-source-id: 8b5afe8262f7d899c05bda438667469454fac5af