Avoid the quadratic case when checking subtyping between one union that contains another
Summary:
Given an enum union `T = 'a' | 'b' | 'c' | ...` and a union `U = T | void`, when flowing `T -> U`, we unfortunately do not hit the `n log n` enum case when checking whether `T` is a subset of `U`, and instead are forced to flatten both unions and perform that quadratic check that each element of `T` is an element of `U`.

This is unnecessary, since `U` is by definition a superset of of `T`, it just falls through our optimizations because we had not considered the case where the upper bound union actually contained the lower bound union in its definition. This adds another `n log n` check to the union optimization guard before the quadratic check if `U` contains any union types.

Note that `n log n` is not the best we can do here. `U` contains `T` nominally, so we should in theory be able to do a constant time check here if we give each union an id. The question is whether the time benefit outweighs the additional memory usage, and whether adding more ids would increase perturbation of recheck optimizations.

Reviewed By: avikchaudhuri

Differential Revision: D15340611

fbshipit-source-id: 4ce94740619de132c16128407ad5abf1c4ce51cc