[libs] Codemod all inexact object types to use explicit syntax
Summary:
This diff used `flow-upgrade` to codemod our libdefs to use explicit inexact objects.

I ran `flow-upgrade 0.83.0 0.99.0` to do it. I used a custom fork of `flow-upgrade`, since it's currently broken in master. I'm going to follow up with a fix to that as well.

I had to re-record the tests that use the libdefs. note that recast re-formatted a large portion of the libdefs.

Reviewed By: gabelevi

Differential Revision: D15503065

fbshipit-source-id: ffc08be4794befef1eaa9d15dc1292437d493a9a