improving flow_use_op
Summary:
Background:
The following test generates a badly positioned error message.
  type T = <A>((A) => mixed) => (A & A);
  type fn = (arg: string) => number;
  type arg = $Call<T, fn>;
  const t: arg = 5;
We would expect the use of arg at line 4 to be blamed.
Flow, instead, blames the introduction of arg at line 3.

Changes:
Given a use type and an alternative use_op, the function flow_use_op decides whether to replace the original use_op of the type with the alternative one.
In this diff, TypeApplication is now in the group that is supposed to be replaced.
A little fiddling is made to make the function a bit more readable and efficient.

No other changes are observed when running runtest.sh.
Still, this function is creepy and should be improved.

Reviewed By: avikchaudhuri

Differential Revision: D15524115

fbshipit-source-id: 201ef35510034ed2d5452b18257f636a698a5fe8