[get-def] use the typed AST instead of the type-tables
Summary:
Compared to other services there's one extra piece of information that the
get-def module requires.

In expressions of the form
```
obj.f  // A
```
when focusing on `f`, we need to keep track of the pair:
```
(type of `obj`, "f")
```
This is later passed to members.ml.

Similarly,
```
import { k } from 'm'; // B
```
corresponded to
```
(type of module "m", "k")
```

This information was populated and stored in the type tables as part of the
main type-checking pass. It is now reconstructed through the TAST pass as
```
type get_def_object_source =
  | GetDefType of Type.t         (* A *)
  | GetDefRequireLoc of ALoc.t   (* B *)

type get_def_member_info = {
  get_def_prop_name: string;
  get_def_object_source: get_def_object_source;
}
```
The require-Loc is looked up through `Context.find_require`.

Reviewed By: nmote

Differential Revision: D15500556

fbshipit-source-id: b245e3dc9342dcee008deeec2bbb21e1ffac64a7