unary constant evaluation in types-first
Summary:
We were careful to ask for annotations on addition and logical operations in signature verification, since Flow has complex logic for these operations, but forgot to do so for unary minus and not, where also the logic is complex.

This diff compensates a bit in the signature generator, but there is still a gap between classic and types-first here. We can leave it as is by design, or do more later.

Reviewed By: panagosg7

Differential Revision: D15565075

fbshipit-source-id: 014a7f05c004489f005bf2e047e4bb86df708023