distinguish classic and types-first merging progress
Summary: The progress message `merged files...` signaled the last step of checking in classic. Using the same message in types-first for something that is *not* the last step, but the penultimate step, can be misleading. This diff changes that message to `merged types of files...` instead; the last step continues to show progress with a message of the form `checking files...`.

Reviewed By: jbrown215

Differential Revision: D15594360

fbshipit-source-id: 1e682af7e49d8d46293c10515aba1aebe62d93c1