Fix file watcher exiting bug
Summary: `watchman#waitpid` was misbehaving when canceled. Instead of resolving to a canceled promise, it was resolving to `unit`. This led the monitor to think that watchman had crashed and that the Flow monitor needed to exit. This would be triggered when the Flow server exited and we tried to cancel the `watchman#waitpid` promise.

Reviewed By: jbrown215

Differential Revision: D15562762

fbshipit-source-id: 948683d0fcf6fa0b44e5888794571eee42f26ac9