[react] ObjT <: ReactAbstractComponentT should not case on call_t itself
Summary:
This fixes the ObjT <: ReactAbstractComponentT case. This will take the defaultProps that are specified on the object and use them to calculate the config.

The tricky part here was that we still need to ensure that the return type of the callable object is a subtype of React.Node. To do that, I flow the callable object to `(...empty) => React.Node`. This should be safe, since `empty` will become a lower bound of every argument in the callable object, and empty is trivially ok as a lower bound of any type.

I use `mixed` for the statics and proto of the function type. In a FunT <: FunT flow these aren't checked, so I don't think the type here matters very much for this case.

Reviewed By: samwgoldman

Differential Revision: D15580351

fbshipit-source-id: 04197284b5935050723bd22cf7b2a05ce1fdedbf