[react] Use defaultProps supplied by Object Components
Summary:
Callable objects can be used as React components, but we didn't have a case for objects in `get_defaults`. This diff fixes that.

However, this still isn't sufficient for the ObjT <: ReactAbstractComponentT check. That case delegates to the callable function, which doesn't have defaultProps as part of the statics. I address that in the next diff.

Reviewed By: samwgoldman

Differential Revision: D15580352

fbshipit-source-id: fe01de5b34df06cb5e1f791e8025ee47f0018b12