[LSP] Return empty responses rather than errors in some circumstances
Summary:
We don't want to spam users with unactionable error messages, so if certain requests error, we'll send empty responses instead. See the comments for more details.

This is the approach that Mostafa Eweda suggested to deal with the unfortunate fact that it's not terribly uncommon for Flow's IDE services to raise exceptions.

Reviewed By: gabelevi

Differential Revision: D15550510

fbshipit-source-id: 05f5eb8e10ac8643afb640fae879ab22250056eb