Consider type parameter defaults in TypeAppVarianceCheckT
Summary:
When we compare type typeapps with the same root, we take a shortcut and only
compare the type arguments according to the polarity of the type parameter.

This shortcut did not properly account for type parameter defaults. This diff
handles the default case by ensuring that all type arguments are constrained.

Reviewed By: jbrown215

Differential Revision: D15675738

fbshipit-source-id: e485070e994535ab3cae4de7994d2a950ad70502