Shared heap for abstract location tables
Summary:
This currently does not lead to the creation or storage of any tables. It does set up the infrastructure to allow it, though.

Future diffs will look for the presence of the abstract-locations flag and will convert the `ALoc.t`s to have keys, as well as populating the tables.

As a (maybe premature) optimization, we only store the table in the shared heap when we actually need to create a table. This allows us to avoid storing additional information in the shared heap when abstract locations are not turned on.

Reviewed By: avikchaudhuri

Differential Revision: D15440666

fbshipit-source-id: a2d5edd6655f8ccea2147f91032836601f9f568a