Store sig file sigs with ALoc.ts
Summary:
Later in this stack, we start assigning keys to the `ALoc.t`s in the sig AST, and populating the associated tables. Because the locations in the Sig File_sig enter the type system, they have to match up with the locations in the AST.

Therefore, we have to generate the Sig File_sig from the abstractified sig AST, rather than just converting the concrete `Loc.t`s to `ALoc.t`s whenever we feel like it, like we do right now.

As of this diff, we build the Sig File_sig out of the sig AST, after it has been converted to contain `ALoc.t`s.

Reviewed By: avikchaudhuri

Differential Revision: D15456386

fbshipit-source-id: ddfb50dbdafc72d0bb2c1b2544ee894e9f4fd20a