fix .flow + haste weirdness
Summary: Previously, the semantics of .flow files with implicit providesModule were pretty weird. This caused spooky effects, e.g. when a .flow file was provided without providesModule or name reducers or an implementation file, it would not be considered a provider of a Haste module, but as soon as an implementation file with providesModule was provided, it would become one. To "fix" this, we over-conservatively considered all .flow files to have changed signatures. This caused spurious rechecks.

Reviewed By: samwgoldman

Differential Revision: D15336348

fbshipit-source-id: feb8af54f6c67d6290fcc54d88ced4b8fab1c865