Add extra type param to Ast.Identifier.t
Summary:
The previous definition for `Ast.Identifier`:
```
and Identifier : sig
  type 'M t = 'M * 'M t'

  and 'M t' = {
    name: string;
    comments: ('M, unit) Syntax.t option;
  }
end = Identifier
```
was problematic in that it pinned the `'M` parameter of `comments` to the identifier's `'M`.
This meant that for all the typed identifiers we would also have to type the comments
that went with it. This led to a number of calls to `map_ident` and `map_ident_new_type`
to be sprinkled all over statement.ml.

Changing the definition to
```

and Identifier : sig
  type ('M, 'T) t = 'T * 'M t'

  and 'M t' = {
    name: string;
    comments: ('M, unit) Syntax.t option;
  }
end = Identifier
```
the typed part `'T` of the identifier doesn't touch the `comments` part anymore.

The rest of the diff propagates this change.
* `'M Identifier.t` becomes `('M, 'M) Identifier.t`
* `'T Identifier.t` becomes `('M, 'T) Identifier.t`

Reviewed By: samwgoldman

Differential Revision: D15480298

fbshipit-source-id: cf607e630d1bcbe7f252941003c8c561fc816fb3