[trusted-types] Add infrastructure to the context to support trust inference
Summary: Adds a `trust_graph` field to the context to track constraints on trust variables. Also creates a definition for constraints on trust variables, based on `constraint.ml` for constraints on tvars, but simpler: instead of the bounds of a variable containing upper- and lower-bound types (in addition to upper and lower variables), the bounds of a trust variable just contains its current `trust` value.

Reviewed By: dsainati1

Differential Revision: D14610542

fbshipit-source-id: cd9f15a8179baff49b3f6e2fdfab4fa89e479ce6