[trusted-types] Core implementation of trust inference
Summary:
This diff provides the core implementation of trust inference. It's still unused, because trust variables are never created, but it contains the logic for propagating trust across the trust graph. The main work in this diff is in the `trust_checking` module, which now handles both fixed trust and trust variables. I've tried to add detailed documentation about the algorithms used, also discussed briefly here.

As implemented in D14610542, trust variables point to nodes in the trust graph which contain a mutable trust data representing the currently known trust of the variable, and a set of upper and lower bound variables.

* When a type with fixed trust (e.g. `any`) flows into a type with inferred trust, the trust of the lower bound is propagated on the trust graph to the trust variable of the upper bound, and from there to its upper bound types.
* Vice versa for when an inferred trust type flows into a fixed trust type.
* When a type with inferred trust flows into another type with inferred trust, we flow the trust variable of the lower bound to the trust variable of the upper bound, e.g. `X ~> Y`. When this occurs, we add `Y ∪ upper_bound_variables(Y)` to `X`'s upper bound variables and to each of `X`'s lower bound variables's upper bound variables. This maintains the invariant that the upper bounds of a trust variable are closed: if `X ~> Y ~> Z`, then `Z` is in `X`'s upper bound variables. We also flow the current trust of the lower bound into the upper bound and vice versa.

Reviewed By: dsainati1

Differential Revision: D14614055

fbshipit-source-id: 738106177fc01340b49b8d8d3f8b25c51ab330df