Implement “go to type definition” for Hack
Summary:
Implement “go to type definition” for Hack

Currently Hack supports functionality for "Go to definition", which given a statement in the language, `hh_client` will point to the file and location within the file where that statement is defined.

This adds a similar functionality: given a statement in the language, `hh_client` can now point to the file and location within the file where the *type* of the statement is defined.

Reviewed By: pittsw

Differential Revision: D15655333

fbshipit-source-id: f82b882833e74f17588371e5d36cff896f4a1d61