[PR] [LSP] fix range in completion response
Summary:
the range in a TextEdit needs to point to the whole range to replace. when we send back a TextEdit in a completion response, it was just the cursor position (ish?). So if you're completing `foo.ba|` and we return `bar()`, the range needs to cover all of `ba` so that it gets replaced.

Fixes https://github.com/flowtype/flow-for-vscode/issues/335

Pull Request resolved: https://github.com/facebook/flow/pull/7816

Reviewed By: gabelevi

Differential Revision: D15738755

Pulled By: mroch

fbshipit-source-id: 5468cf765ad4a06ce2c0f1597abe1149d28c2dcc