Track why we're rechecking and why we're sending errors to lsp
Summary:
We're pretty good with our logging of user interactions that are initiated by clear lsp commands, like `textDocument/completion`. We see when that request comes in and we see when we respond to it. We're much worse with logging when we show the user errors. One of the main reasons that this is hard is that diagnostics are push-based, and often triggered by filesystem changes.

I plan to improve that. And the first step is keeping track of why we're doing rechecks and why we're sending errors to the persistent connection. This can help us track how well we respond to the user saving 1 file, versus responding to a rebase.

One thing to keep in mind is that since we often combine together multiple rechecks into a single recheck, we might have multiple reasons for doing that single recheck.

Reviewed By: avikchaudhuri

Differential Revision: D15287913

fbshipit-source-id: 5548db4e93506d8f627cd54d5ec44b8001759810