Roughly track unsaved buffers in lsp
Summary:
For some logging I'm building, I want to know whether any open buffers are unsaved. This is a rough approximation. It works like so:

1. When you open a buffer we assume it's saved
2. When we get a didChange, we assume it's unsaved
3. When we get a didSave, we assume it's saved again

This is not perfect. If you make changes and undo them, we'll still treat the buffer as unsaved. I imagine that we probably should hash the contents, remember the on-disk hash, and use that to decide if a buffer is saved. But for now this suffices.

Reviewed By: jbrown215

Differential Revision: D15545140

fbshipit-source-id: d2f723f494fbd520cd71b4cad732130f21f7d938