Utility to lazily get the appropriate ALoc table from an ALoc.t
Summary:
This utility is used in D14091206 to more conveniently convert `ALoc.t`s back to `Loc.t`s.

Ideally we wouldn't do a fetch to the shared heap for each `ALoc.t`, and would instead batch the requests per file, but the way the errors are organized, that's not really straightforward. We could add another cache if this proves to be a bottleneck, but I think the shared heap already has its own cache.

Anyway, since this is lazy, no fetches are done unless the `ALoc.t` is actually represented by a key. Without the abstract locations flag, none are. I'll have the opportunity to optimize this as needed before launching abstract locations.

I'm not sure if this is the best place for such a utility, but it's definitely convenient to have it co-located with the other AST heap functions.

Reviewed By: avikchaudhuri

Differential Revision: D14091205

fbshipit-source-id: 2be6987bbffba954f9f0a563ff0c245a6f2e524f